requirements.txt
fastapi
uvicorn[standard]
black
google-generativeai
python-dotenv